+++ ChangeLog.md
@@ -1,6 +1,17 @@
 Note: This is an Azure Storage only package. The all up Azure node sdk still has the old storage bits in there. In a future release, those storage bits will be removed and an npm dependency to this storage node sdk will 
 be taken. This is a GA release and the changes described below indicate the changes from the Azure node SDK 0.9.8 available here - https://github.com/Azure/azure-sdk-for-node.
 
+2018.06 Version 2.10.0
+
+ALL
+* Updated storage service version to 2018-03-28.
+
+BLOB
+* Fixed a bug that `DeleteRetentionPolicy.Days` should be `number` instead of `string` when calling `getServiceProperties`.
+* Added a method `getAccountProperties` to `blobService`.
+* Added a method `createBlockFromURL` to `blobService`.
+* Added support for static website service properties (in preview).
+
 2018.05 Version 2.9.0-preview
 
 ALL
@@ -11,7 +22,7 @@ ALL
 BLOB
 * Added `createBlobServiceWithTokenCredential()` to create `BlobService` object with bearer tokens such as OAuth access token (in preview).
 * Added support for '$web' as a valid blob container name for static website.
-* Added support for write-once read-many containers.
+* Added support for write-once read-many containers (in preview).
 * The `Get Container Properties` and `List Containers` APIs now return two new properties indicating whether the container has an immutability policy or a legal hold.
 * The `Get Blob Properties` and `List Blobs` APIs now return the creation time of the blob as a property.